+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [11.0.7] - 2025-07-10
+
 ### Added
 
 - `footer` to `FormsListStyles`